SearchBox: add center to Stars
`+ refactoring`
