+++ src/components/Map/behaviour/useFeatureMarker.tsx
@@ -61,7 +61,7 @@ export const useFeatureMarker = (map) => {
   useUpdateFeatureMarker(map, feature);
   useUpdatePreviewMarker(map, preview);
 
-  // hide the icon when tiles are fetched TODO sometimes broken (zoom problem) (also maybe causes webgl blackout)
+  // hide the icon when tiles are fetched TODO sometimes broken (zoom problem)
   useEffect(() => {
     if (map) {
       const handle = setInterval(() => {
