+++ src/components/SearchBox/AutocompleteInput.tsx
@@ -3,10 +3,11 @@ import { Autocomplete, InputBase } from '@mui/material';
 import { useFeatureContext } from '../utils/FeatureContext';
 import { renderOptionFactory, buildPhotonAddress } from './renderOptionFactory';
 import { t } from '../../services/intl';
-import { onHighlightFactory, onSelectedFactory } from './onSelectedFactory';
+import { onSelectedFactory } from './onSelectedFactory';
 import { useMobileMode } from '../helpers';
 import { useUserThemeContext } from '../../helpers/theme';
 import { useMapStateContext } from '../utils/MapStateContext';
+import { onHighlightFactory } from './onHighlightFactory';
 
 const useFocusOnSlash = () => {
   const inputRef = React.useRef<HTMLInputElement>(null);
@@ -104,13 +105,7 @@ export const AutocompleteInput = ({
           autocompleteRef={autocompleteRef}
         />
       )}
-      // renderOption={renderOptionFactory(inputValue, currentTheme)}
-      renderOption={(props, option) => (
-        // eslint-disable-next-line react/jsx-props-no-spreading
-        <li {...props}>
-          {renderOptionFactory(inputValue, currentTheme)(props, option)}
-        </li>
-      )}
+      renderOption={renderOptionFactory(inputValue, currentTheme)}
     />
   );
 };
