+++ src/components/SearchBox/renderOptionFactory.tsx
@@ -1,103 +1,13 @@
 import React from 'react';
-import styled from 'styled-components';
-
-import StarIcon from '@mui/icons-material/Star';
-import FolderIcon from '@mui/icons-material/Folder';
-import SearchIcon from '@mui/icons-material/Search';
 import { Grid, Typography } from '@mui/material';
-import { useMapStateContext } from '../utils/MapStateContext';
 import Maki from '../utils/Maki';
 import { highlightText } from './highlightText';
 import { join } from '../../utils';
 import { getPoiClass } from '../../services/getPoiClass';
-import { t } from '../../services/intl';
-
-/** photon
-{
-  "features": [
-    {
-      "geometry": {
-        "coordinates": [
-          16.5920871,
-          49.2416882
-        ],
-        "type": "Point"
-      },
-      "type": "Feature",
-      "properties": {
-        "osm_id": 2493045013,
-        "country": "Česko",
-        "city": "Brno",
-        "countrycode": "CZ",
-        "postcode": "612 00",
-        "county": "Jihomoravský kraj",
-        "type": "house",
-        "osm_type": "N",
-        "osm_key": "leisure",
-        "street": "Podhájí",
-        "district": "Řečkovice",
-        "osm_value": "sports_centre",
-        "name": "VSK MENDELU",
-        "state": "Jihovýchod"
-      }
-    },
-    // housenumber
-    {
-      geometry: { coordinates: [14.4036424, 50.098012], type: 'Point' },
-      type: 'Feature',
-      properties: {
-        osm_id: 296816783,
-        country: 'Czechia',
-        city: 'Prague',
-        countrycode: 'CZ',
-        postcode: '16000',
-        type: 'house',
-        osm_type: 'N',
-        osm_key: 'place',
-        housenumber: '8',
-        street: 'Dejvická',
-        district: 'Dejvice',
-        osm_value: 'house',
-        state: 'Prague',
-      },
-    }
-  ]
- */
-
-const IconPart = styled.div`
-  width: 50px;
-  text-align: center;
-  padding-right: 10px;
-  font-size: 10px;
-  color: ${({ theme }) => theme.palette.text.secondary};
-`;
-
-const getDistance = (point1, point2) => {
-  const lat1 = (parseFloat(point1.lat) * Math.PI) / 180;
-  const lng1 = (parseFloat(point1.lon) * Math.PI) / 180;
-  const lat2 = (parseFloat(point2.lat) * Math.PI) / 180;
-  const lng2 = (parseFloat(point2.lon) * Math.PI) / 180;
-  const latdiff = lat2 - lat1;
-  const lngdiff = lng2 - lng1;
-
-  return (
-    6372795 *
-    2 *
-    Math.asin(
-      Math.sqrt(
-        Math.sin(latdiff / 2) ** 2 +
-          Math.cos(lat1) * Math.cos(lat2) * Math.sin(lngdiff / 2) ** 2,
-      ),
-    )
-  );
-};
-
-const useMapCenter = () => {
-  const {
-    view: [, lat, lon],
-  } = useMapStateContext();
-  return { lon, lat };
-};
+import { renderOverpass } from './options/overpass';
+import { renderPreset } from './options/preset';
+import { getDistance, IconPart, renderLoader, useMapCenter } from './utils';
+import { renderStar } from './options/stars';
 
 const getAdditionalText = (props) => {
   const address = [
@@ -119,109 +29,111 @@ export const buildPhotonAddress = ({
   streetnumber: snum,
 }) => join(street ?? place ?? city, ' ', hnum ? hnum.replace(' ', '/') : snum);
 
+/** photon
+[
+  {
+    geometry: {
+      coordinates: [16.5920871, 49.2416882],
+      type: 'Point',
+    },
+    type: 'Feature',
+    properties: {
+      osm_id: 2493045013,
+      country: 'Česko',
+      city: 'Brno',
+      countrycode: 'CZ',
+      postcode: '612 00',
+      county: 'Jihomoravský kraj',
+      type: 'house',
+      osm_type: 'N',
+      osm_key: 'leisure',
+      street: 'Podhájí',
+      district: 'Řečkovice',
+      osm_value: 'sports_centre',
+      name: 'VSK MENDELU',
+      state: 'Jihovýchod',
+    },
+  },
+  // housenumber
+  {
+    geometry: { coordinates: [14.4036424, 50.098012], type: 'Point' },
+    type: 'Feature',
+    properties: {
+      osm_id: 296816783,
+      country: 'Czechia',
+      city: 'Prague',
+      countrycode: 'CZ',
+      postcode: '16000',
+      type: 'house',
+      osm_type: 'N',
+      osm_key: 'place',
+      housenumber: '8',
+      street: 'Dejvická',
+      district: 'Dejvice',
+      osm_value: 'house',
+      state: 'Prague',
+    },
+  },
+];
+ */
+const renderGeocoder = (option, currentTheme, inputValue) => {
+  const { geometry, properties } = option;
+  const { name, osm_key: tagKey, osm_value: tagValue } = properties;
+
+  const [lon, lat] = geometry.coordinates;
+  const mapCenter = useMapCenter();
+  const dist = getDistance(mapCenter, { lon, lat }) / 1000;
+  const distKm = dist < 10 ? Math.round(dist * 10) / 10 : Math.round(dist); // TODO save imperial to mapState and multiply 0.621371192
+
+  const text = name || buildPhotonAddress(properties);
+  const additionalText = getAdditionalText(properties);
+  const poiClass = getPoiClass({ [tagKey]: tagValue });
+
+  return (
+    <>
+      <IconPart>
+        <Maki
+          ico={poiClass.class}
+          style={{ width: '20px', height: '20px', opacity: 0.5 }}
+          title={`${tagKey}=${tagValue}`}
+          invert={currentTheme === 'dark'}
+        />
+        <div>{distKm} km</div>
+      </IconPart>
+      <Grid item xs>
+        {highlightText(text, inputValue)}
+        <Typography variant="body2" color="textSecondary">
+          {additionalText}
+        </Typography>
+      </Grid>
+    </>
+  );
+};
+
+const renderOption = (inputValue, currentTheme, option) => {
+  const { preset, overpass, star, loader } = option;
+  if (overpass) {
+    return renderOverpass(overpass);
+  }
+
+  if (star) {
+    return renderStar(star);
+  }
+
+  if (loader) {
+    return renderLoader();
+  }
+
+  if (preset) {
+    return renderPreset(preset, inputValue);
+  }
+
+  return renderGeocoder(option, currentTheme, inputValue);
+};
+
 export const renderOptionFactory =
-  (inputValue, currentTheme) => (_props, option) => {
-    const { geometry, preset, properties, overpass, star, loader } = option;
-    if (overpass) {
-      return (
-        <>
-          <IconPart>
-            <SearchIcon />
-          </IconPart>
-          <Grid item xs>
-            <span style={{ fontWeight: 700 }}>{overpass.label}</span>
-            <Typography variant="body2" color="textSecondary">
-              overpass search
-            </Typography>
-          </Grid>
-        </>
-      );
-    }
-
-    if (star) {
-      return (
-        <>
-          <IconPart>
-            <StarIcon />
-          </IconPart>
-          <Grid item xs={8}>
-            <span style={{ fontWeight: 700 }}>{star.label}</span>
-            <Typography variant="body2" color="textSecondary">
-              {star.poiType}
-            </Typography>
-          </Grid>
-        </>
-      );
-    }
-
-    if (loader) {
-      return (
-        <>
-          <IconPart />
-          <Grid item xs>
-            <Typography>
-              {t('loading')}
-              <span className="dotloader">.</span>
-              <span className="dotloader">.</span>
-              <span className="dotloader">.</span>
-            </Typography>
-          </Grid>
-        </>
-      );
-    }
-
-    if (preset) {
-      const { name } = preset.presetForSearch;
-      const additionalText =
-        preset.name === 0
-          ? ` (${preset.presetForSearch.texts.find(
-              (_, idx) => preset.textsByOne[idx] > 0,
-            )}…)`
-          : '';
-
-      return (
-        <>
-          <IconPart>
-            <FolderIcon />
-          </IconPart>
-          <Grid item xs>
-            {highlightText(`${name}${additionalText}`, inputValue)}
-            <Typography variant="body2" color="textSecondary">
-              {t('searchbox.category')}
-            </Typography>
-          </Grid>
-        </>
-      );
-    }
-
-    const { name, osm_key: tagKey, osm_value: tagValue } = properties; // @TODO
-
-    const [lon, lat] = geometry.coordinates;
-    const mapCenter = useMapCenter();
-    const dist = getDistance(mapCenter, { lon, lat }) / 1000;
-    const distKm = dist < 10 ? Math.round(dist * 10) / 10 : Math.round(dist); // TODO save imperial to mapState and multiply 0.621371192
-
-    const text = name || buildPhotonAddress(properties);
-    const additionalText = getAdditionalText(properties);
-    const poiClass = getPoiClass({ [tagKey]: tagValue });
-
-    return (
-      <>
-        <IconPart>
-          <Maki
-            ico={poiClass.class}
-            style={{ width: '20px', height: '20px', opacity: 0.5 }}
-            title={`${tagKey}=${tagValue}`}
-            invert={currentTheme === 'dark'}
-          />
-          <div>{distKm} km</div>
-        </IconPart>
-        <Grid item xs>
-          {highlightText(text, inputValue)}
-          <Typography variant="body2" color="textSecondary">
-            {additionalText}
-          </Typography>
-        </Grid>
-      </>
+  (inputValue, currentTheme) => (props, option) =>
+    (
+      // eslint-disable-next-line react/jsx-props-no-spreading
+      <li {...props}>{renderOption(inputValue, currentTheme, option)}</li>
     );
-  };
