+++ src/components/utils/StarsContext.tsx
@@ -2,9 +2,15 @@ import React, { createContext, useContext } from 'react';
 import { usePersistedState } from './usePersistedState';
 import { useFeatureContext } from './FeatureContext';
 import { getShortId } from '../../services/helpers';
-import { getLabel, getPoiType } from '../../helpers/featureLabel';
+import { getLabel, getHumanPoiType } from '../../helpers/featureLabel';
+import { LonLat } from '../../services/types';
 
-type Star = { shortId: string; poiType: string; label: string };
+export type Star = {
+  shortId: string;
+  poiType: string;
+  label: string;
+  center: LonLat;
+};
 
 interface StarsType {
   stars: Star[];
@@ -19,10 +25,11 @@ const hasStar = (stars: Star[], shortId: string) =>
 
 export const StarsProvider = ({ children }) => {
   const { feature } = useFeatureContext();
+  const shortId = feature ? getShortId(feature.osmMeta) : undefined;
 
   const [stars, setStars] = usePersistedState<Star[]>('stars', []);
-  const shortId = feature ? getShortId(feature.osmMeta) : undefined;
   const isStarred = hasStar(stars, shortId);
+
   const toggleStar = () => {
     if (!shortId) {
       return;
@@ -33,8 +40,9 @@ export const StarsProvider = ({ children }) => {
         ? data.filter((star) => star.shortId !== shortId)
         : data.concat({
             shortId,
-            poiType: getPoiType(feature),
+            poiType: getHumanPoiType(feature),
             label: getLabel(feature),
+            center: feature.center,
           }),
     );
   };
